pyproject.toml
[project]
name = "whatflix-auth-api"
version = "0.1.0"
description = "Auth API for Whatflix"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.0",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0"
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0"
]
